Move CustomItemButton to utils/minecraft/ItemButton
